avformat/dvd: new dvd:// protocol for reading dvd folder/images
dvd:// protocol uses libdvdread for opening folders and/or disc-images,
it then either identifies the longest title-set, or uses the title-set
parameter.

After opening the dvd, libdvdread-protocol will read and output all
VOBs, in the cell and pack sequence defined by the program-chain.

The (split)-VOB content that is streamed from the folder/disc by
libdvdread, is given to ffmpeg probed demuxer/decoder, which decodes
and presents the frames.

The chapter and language stream information is present in the
side-files (BUP/IFO), which are parsed for identifying the correct
main video-title-set, but not attached to the stream information.

Signed-off-by: Stanislav Ionascu <stanislav.ionascu@gmail.com>